K-operator for Parkinson's data
This is an application of the K-operator (Mannone, Fazio, Marwan, 2024 to appear) to investigate neurological diseases from fMRI data. Here it is applied on some selected Parkinson's disease patients and healthy control from the PPMI dataset. Visualization of the connectivity matrices and connectome: nilearn, and Hough et al. Structure and implementation of the K-operator: by M. Mannone. Due to the size of the file, only the essential code is reported here. The same scheme can be applied to all patients.